+++ src/complete.cpp
@@ -7,3 +7,8 @@ void wps::complete( const eosio::time_point_sec voting_period )
     auto state = _state.get();
     auto settings = _settings.get();
 }
+
+void wps::check_completed()
+{
+    check( current_time_point() < _state.get().next_voting_period, "current voting period phase is now completed, any account must now execute [complete] ACTION to continue");
+}